+++ pkg/gateway/server_http.go
@@ -10,7 +10,6 @@ import (
 	"net"
 	"os"
 	"runtime"
-	"runtime/debug"
 	"sync/atomic"
 	"syscall"
 	"time"
@@ -29,7 +28,7 @@ import (
 	"github.com/hertz-contrib/pprof"
 	"github.com/nite-coder/bifrost/internal/pkg/safety"
 	"github.com/nite-coder/bifrost/pkg/config"
-	"github.com/nite-coder/blackbear/pkg/cast"
+	"github.com/nite-coder/bifrost/pkg/zero"
 	prom "github.com/prometheus/client_golang/prometheus"
 	"golang.org/x/sys/unix"
 )
@@ -67,7 +66,6 @@ func newHTTPServer(bifrost *Bifrost, serverOptions config.ServerOptions, tracers
 	hzOpts := []hzconfig.Option{
 		server.WithDisableDefaultDate(true),
 		server.WithDisablePrintRoute(true),
-		server.WithSenseClientDisconnection(true),
 		server.WithTraceLevel(stats.LevelBase),
 		server.WithReadTimeout(time.Second * 60),
 		server.WithWriteTimeout(time.Second * 60),
@@ -154,7 +152,17 @@ func newHTTPServer(bifrost *Bifrost, serverOptions config.ServerOptions, tracers
 			},
 		}
 
-		listener, err := bifrost.zero.Listener(ctx, "tcp", serverOptions.Bind, listenerConfig)
+		listenerOptions := &zero.ListenerOptions{
+			Network: "tcp",
+			Address: serverOptions.Bind,
+			Config:  listenerConfig,
+		}
+
+		if serverOptions.ProxyProtocol {
+			listenerOptions.ProxyProtocol = true
+		}
+
+		listener, err := bifrost.zero.Listener(ctx, listenerOptions)
 		if err != nil {
 			return nil, err
 		}
@@ -266,11 +274,13 @@ func newHTTPServer(bifrost *Bifrost, serverOptions config.ServerOptions, tracers
 		}
 		tlsConfig.Certificates = append(tlsConfig.Certificates, cert)
 		hzOpts = append(hzOpts, server.WithTLS(tlsConfig))
+	} else {
+		hzOpts = append(hzOpts, server.WithSenseClientDisconnection(true))
 	}
 
 	httpServer.options = &serverOptions
 
-	h := server.Default(hzOpts...)
+	h := server.New(hzOpts...)
 
 	if len(serverOptions.RemoteIPHeaders) > 0 || len(serverOptions.TrustedCIDRS) > 0 {
 		clientIPOptions := app.ClientIPOptions{
@@ -335,18 +345,6 @@ func newHTTPServer(bifrost *Bifrost, serverOptions config.ServerOptions, tracers
 }
 
 func (s *HTTPServer) Run() {
-	s.server.PanicHandler = func(ctx context.Context, c *app.RequestContext) {
-		stackTrace := cast.B2S(debug.Stack())
-		err := c.Errors.Last()
-		slog.Error("http request panic recovered",
-			slog.String("server_id", s.options.ID),
-			slog.String("uri", cast.B2S(c.Request.RequestURI())),
-			slog.String("method", cast.B2S(c.Request.Method())),
-			slog.String("stack", stackTrace),
-			slog.Any("error", err),
-		)
-	}
-
 	slog.Info("starting server", "id", s.options.ID, "bind", s.options.Bind, "transporter", s.server.GetTransporterName())
 	_ = s.server.Run()
 }