+++ pkg/tracer/prometheus/tracer.go
@@ -33,8 +33,13 @@ func genRequestDurationLabels(c *app.RequestContext) prom.Labels {
 	labels[labelServerID] = defaultValIfEmpty(serverID, unknownLabelValue)
 	labels[labelRouteID] = defaultValIfEmpty(routeID, unknownLabelValue)
 	labels[labelServiceID] = defaultValIfEmpty(serviceID, unknownLabelValue)
-	labels[labelMethod] = defaultValIfEmpty(string(c.Request.Method()), unknownLabelValue)
 	labels[labelStatusCode] = defaultValIfEmpty(strconv.Itoa(c.Response.Header.StatusCode()), unknownLabelValue)
+
+	method := variable.GetString(variable.HTTPRequestMethod, c)
+	if method == "" {
+		method = cast.B2S(c.Method())
+	}
+	labels[labelMethod] = defaultValIfEmpty(method, unknownLabelValue)
 
 	path := variable.GetString(variable.HTTPRoute, c)
 	if path == "" {